+++ src/Column/ColumnDefinitionBuilder.php
@@ -4,13 +4,16 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Oracle\Column;
 
+use LogicException;
 use Yiisoft\Db\Constant\ColumnType;
 use Yiisoft\Db\Constant\ReferentialAction;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\QueryBuilder\AbstractColumnDefinitionBuilder;
 use Yiisoft\Db\Schema\Column\ColumnInterface;
+use Yiisoft\Db\Schema\Column\EnumColumn;
 
 use function ceil;
+use function in_array;
 use function log10;
 use function strtoupper;
 
@@ -61,23 +64,20 @@ final class ColumnDefinitionBuilder extends AbstractColumnDefinitionBuilder
 
         if (empty($check)) {
             $name = $column->getName();
-
-            if (empty($name)) {
-                return '';
+            if (!empty($name)) {
+                $type = $column->getType();
+                if (in_array($type, [ColumnType::ARRAY, ColumnType::STRUCTURED, ColumnType::JSON], true)) {
+                    return version_compare($this->queryBuilder->getServerInfo()->getVersion(), '21', '<')
+                        ? ' CHECK (' . $this->queryBuilder->getQuoter()->quoteSimpleColumnName($name) . ' IS JSON)'
+                        : '';
+                }
+                if ($type === ColumnType::BOOLEAN) {
+                    return ' CHECK (' . $this->queryBuilder->getQuoter()->quoteSimpleColumnName($name) . ' IN (0,1))';
+                }
             }
-
-            return match ($column->getType()) {
-                ColumnType::ARRAY, ColumnType::STRUCTURED, ColumnType::JSON
-                    => version_compare($this->queryBuilder->getServerInfo()->getVersion(), '21', '<')
-                    ? ' CHECK (' . $this->queryBuilder->getQuoter()->quoteSimpleColumnName($name) . ' IS JSON)'
-                    : '',
-                ColumnType::BOOLEAN
-                    => ' CHECK (' . $this->queryBuilder->getQuoter()->quoteSimpleColumnName($name) . ' IN (0,1))',
-                default => '',
-            };
         }
 
-        return " CHECK ($check)";
+        return parent::buildCheck($column);
     }
 
     protected function buildOnDelete(string $onDelete): string
@@ -133,6 +133,7 @@ final class ColumnDefinitionBuilder extends AbstractColumnDefinitionBuilder
                     => version_compare($this->queryBuilder->getServerInfo()->getVersion(), '21', '>=')
                     ? 'json'
                     : 'clob',
+                ColumnType::ENUM => 'varchar2(' . $this->calcEnumSize($column) . ' BYTE)',
                 default => 'varchar2',
             },
             'timestamp with time zone' => 'timestamp' . ($size !== null ? "($size)" : '') . ' with time zone',
@@ -145,5 +146,24 @@ final class ColumnDefinitionBuilder extends AbstractColumnDefinitionBuilder
     protected function getDefaultUuidExpression(): string
     {
         return 'sys_guid()';
+    }
+
+    private function calcEnumSize(ColumnInterface $column): int
+    {
+        $size = $column->getSize();
+        if ($size !== null) {
+            return $size;
+        }
+
+        if ($column instanceof EnumColumn) {
+            return max(
+                array_map(
+                    strlen(...),
+                    $column->getValues(),
+                ),
+            );
+        }
+
+        throw new LogicException('Cannot calculate enum size. Set the size explicitly or use `EnumColumn` instance.');
     }
 }